reference/langchain-qs/requirements.txt
langchain>=0.3.26
langchain-community>=0.3.27
langchain-core>=0.3.29
langgraph>=0.4.1
langgraph-swarm>=0.1.0
pymongo>=4.10.1
langchain-mongodb>=0.2.0
langgraph-checkpoint-mongodb>=0.1.3
langchain-aws>=0.2.2
boto3>=1.35.40
streamlit>=1.39.0
streamlit-extras>=0.4.0
python-dotenv>=1.0.1
watchdog>=4.0.0
